[TASK] custom service providers
